Filtering of labels from the generated config
Adds a new argument `-config.drop-labels` which is a comma-separated
list of labels you want to drop from the generated configuration.

i.e. `-config.drop-labels container_arn` will remove `container_arn`.

Annoying that it requires `reflect` but the alternative is an `if/then`
for each field that we might want to remove.